pyproject.toml
[tool.poetry]
name = "document-complexity"
version = "0.1.0"
description = ""
authors = ["nandar <nandarkhine14380@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.47.0"
pdfplumber = "^0.11.7"
pymupdf = "^1.26.3"
numpy = "^2.3.1"
python-docx = "^1.2.0"
python-pptx = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
